This branch provides a plugin version for slick 3. This is a bit of a slash-and-burn hack to get things up and running. I’ve removed all of the slick.Database creation and left only the evolutions plugin.
Due to Slick re-naming of the top-level package from scala.slick to just slick, there was a problem with the way database packages were being looked up in the config file: the config key name “slick” has ended up polluted with slick library config. Because of this, I have changed the config file key to be “slick3”.

I had to comment out an assertion in the ReflectionUtils that was blowing up on something that has changed in, I think, the main Table class. So far the result is benign (!).

This gets us up-and-running, at least, until a better, “official” release comes along.